Fixed the problem with %p parameters in 64 bits.
The problem is caused because TMF/TMC annotations in PDB files identify %p parameters as ItemPtr in both 32 and 64 bit.
So, we have no way to differentiate a 32 bit %p from a 64 bit one.
For the time being we have no clean solution, the current solution is to search the PDB for well known pointer types such as PVOID and check their sizes.
When 64 bit pointer types are found, %p parameters are converted to %016I64X, that is an uppercase hexadecimeal 16 character string padded with zeros.

Only basic tests performed, further testing is needed.